sorting by num proteins in single endpoint queries + minor changes in granular counters
